Updated plots for Scenario 1.3, 1.4 and 1.2
The plots from Scenario 2.1 would need to be fixed. The matlab script used to generate these plots are in the file: \GitHub\FeatureExtraction_SafetyMetrics_SafetyMetricsClass\script_AVConsistency_processSceData_sce1_2.m

Problem: The length of the track with the detour is likely less than the distance being plotted. The simulation results are located in \GitHub\TrafficSimulators_Project_SUMOSimulationForADSProject\Scenarios\Scenario1_2\res